+++ subgraphs.js
@@ -2,7 +2,7 @@ import express from 'express';
 import http from 'http';
 import cors from 'cors';
 import bodyParser from 'body-parser';
-import { getSchema as getUsersSchema } from "@apollosolutions/retail-supergraph-users/server.js";
+import { getSchema as getUsersSchema } from "@apollosolutions/retail-supergraph-users/subgraph.js";
 import {expressMiddleware} from "@apollo/server/express4";
 import {ApolloServer} from "@apollo/server";
 import { ApolloServerPluginDrainHttpServer } from '@apollo/server/plugin/drainHttpServer';
